Attempt using resin.io with docker-compose
I do not expect this attempt to work, because of the device mapping I had to do from the host system when we were only going one Docker deep.

Now that it's Docker-in-Docker inception, I think I'll need to map the devices from the top-level in to the host Docker.  Not sure where to do that yet, since Resin is the one starting that process...